Extract P2P replication tests from production module
